[Refactoring ⚙️] useTransformBodydata를 utils로 분리
