+++ src/hooks/query/useGetFetchRecentBodyData.ts
@@ -2,7 +2,7 @@ import { useQuery } from "@tanstack/react-query"
 
 import bodyDataAPI from "@apis/domain/bodydata"
 
-import { transformBodyData } from "../useTransformBodyData"
+import { transformBodyData } from "@utils/transBodyData"
 
 export const useGetFetchRecentData = () => {
   const {
